# Speed up serialization and deserializer.