--- a/EEUniverse.Library.Tests/Program.cs
+++ b/EEUniverse.Library.Tests/Program.cs
@@ -13,34 +13,6 @@ namespace EEUniverseLibrary.Tests
     {
         public static void Main(string[] args)
         {
-			var msg = new Message
-			(
-				ConnectionScope.Lobby,
-				MessageType.ChatOld,
-				"str",
-				12345,
-				-12345,
-				123456.0d,
-				true,
-				false,
-				false,
-				new byte[] { 1, 2, 3 },
-				new MessageObject()
-					.Add("a", "str")
-					.Add("b", 12345)
-					.Add("c", -12345)
-					.Add("d", 123456.0d)
-					.Add("e", true)
-					.Add("f", false)
-					.Add("g", new byte[] { 1, 2, 3 })
-			);
-
-			var bytes = Serializer.Serialize(msg);
-
-			var d1 = Serializer.Deserialize(bytes);
-			var d2 = Serializer.Deserialize(new System.ReadOnlySpan<byte>(bytes));
-
-			System.Console.WriteLine("done");
         }
     }
 }