--- a/EEUniverse.Library/EEUniverse.Library.csproj
+++ b/EEUniverse.Library/EEUniverse.Library.csproj
@@ -29,4 +29,8 @@
     <DebugSymbols>true</DebugSymbols>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.6.0" />
+  </ItemGroup>
+
 </Project>