The current serializer/deserializer is a bit slow, this commit speeds things up.

``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 8.1 (6.3.9600.0)
Unknown processor
Frequency=2924198 Hz, Resolution=341.9741 ns, Timer=TSC
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT


```
|                       Method |        Mean |      Error |     StdDev |
|----------------------------- |------------:|-----------:|-----------:|
|    Complicated_Serialize_Old | 1,579.55 ns | 18.2802 ns | 17.0993 ns |
|   Complicated_Serialize_Fast | 1,083.86 ns |  9.4249 ns |  8.3550 ns |
|  Complicated_Deserialize_Old | 1,968.07 ns | 18.4542 ns | 15.4101 ns |
| Complicated_Deserialize_Fast | 1,340.28 ns | 12.7395 ns | 11.9165 ns |
|        Nothing_Serialize_Old |   154.25 ns |  2.0295 ns |  1.7991 ns |
|       Nothing_Serialize_Fast |    26.41 ns |  0.2338 ns |  0.2073 ns |
|      Nothing_Deserialize_Old |   145.48 ns |  1.7362 ns |  1.6240 ns |
|     Nothing_Deserialize_Fast |    72.66 ns |  0.9093 ns |  0.8505 ns |


Benchmark code: (notice: if you'd want to get this to work, you'd have to go to about the 8th/9th commit, somewhere around there, and get `BenchmarkDotNet`)

```cs
// Ideally these would be automated tests with a mocked server.
// Potential libraries to use:
//  - Specflow (https://github.com/techtalk/SpecFlow); To write human-readable acceptance tests.
//  - Moq (https://github.com/moq/moq4); To easily mock functions. (mainly to fake the client part, because connecting to the real Everybody Edits Universe™ servers is a bad idea...)

// For now it's being manually tested...

using BenchmarkDotNet.Attributes;
using EEUniverse.Library;
using System;

namespace EEUniverseLibrary.Tests
{
	public class Program
	{
		private Message _complicated = new Message
		(
			ConnectionScope.Lobby,
			MessageType.ChatOld,
			"str",
			12345,
			-12345,
			123456.0d,
			true,
			false,
			false,
			new byte[] { 1, 2, 3 },
			new MessageObject()
				.Add("a", "str")
				.Add("b", 12345)
				.Add("c", -12345)
				.Add("d", 123456.0d)
				.Add("e", true)
				.Add("f", false)
				.Add("g", new byte[] { 1, 2, 3 })
		);

		private Message _nothing = new Message(ConnectionScope.Lobby, MessageType.ChatOld);

		public Program()
		{
			_complicatedBytes = Serializer.Serialize(_complicated);
			_nothingBytes = Serializer.Serialize(_nothing);
		}

		private byte[] _complicatedBytes;
		private byte[] _nothingBytes;

		[Benchmark]
		public byte[] Complicated_Serialize_Old()
		{
			return Serializer.Serialize(_complicated);
		}

		[Benchmark]
		public Span<byte> Complicated_Serialize_Fast()
		{
			return Serializer.SerializeFast(_complicated);
		}

		[Benchmark]
		public Message Complicated_Deserialize_Old()
		{
			return Serializer.Deserialize(_complicatedBytes);
		}

		[Benchmark]
		public Message Complicated_Deserialize_Fast()
		{
			return Serializer.Deserialize(new ReadOnlySpan<byte>(_complicatedBytes));
		}

		[Benchmark]
		public byte[] Nothing_Serialize_Old()
		{
			return Serializer.Serialize(_nothing);
		}

		[Benchmark]
		public Span<byte> Nothing_Serialize_Fast()
		{
			return Serializer.SerializeFast(_nothing);
		}

		[Benchmark]
		public Message Nothing_Deserialize_Old()
		{
			return Serializer.Deserialize(_nothingBytes);
		}

		[Benchmark]
		public Message Nothing_Deserialize_Fast()
		{
			return Serializer.Deserialize(new ReadOnlySpan<byte>(_nothingBytes));
		}

		public static void Main(string[] args)
		{
			BenchmarkDotNet.Running.BenchmarkRunner.Run<Program>();
		}
	}
}
```