@EEJesse 2019-10-17
Although greatly appreciated, I don't think it's worth the loss of readability. We're talking about 500 nanoseconds here, for complex messages.

Dependencies are also unwanted, especially "System.Runtime.CompilerServices.Unsafe".

I do like the "IEnumerable<object>" part though. Could you create a new PR with just this change?
(keep the documentation)